twitter_scraper/requirements.txt
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
jmespath==0.10.0
joblib==0.17.0
nltk==3.5
oauthlib==3.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pycld3==0.20
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
textblob==0.15.3
tqdm==4.50.2
tweepy==3.9.0
urllib3==1.25.10
